Update README.md
The Good Doctor series is all about a Dr. Shaun Murphy, a doctor who has autism, and despite to his autism, he is very professional, very serious about his work and sometimes, save lives. Some doctors in the same hospital don't want Shaun to work with them, because sometimes, Shaun does things that very unnecessary things like, scaring people by telling the truth of their sickness and tells that the patient is going to die. Despite his condition, he still reach his dream to be a surgeon and go to the hospital that he really wants and help or even save lives of the patients.